Change teleport cost to double
This doesn't affect item or xp costs, but can be useful when it comes to
economy based costs.

+++ src/main/kotlin/dev/mizarc/waystonewarps/application/services/ConfigService.kt
@@ -7,5 +7,5 @@ interface ConfigService {
     fun getTeleportTimer(): Int
     fun getTeleportCostType(): CostType
     fun getTeleportCostItem(): String
-    fun getTeleportCostAmount(): Int
+    fun getTeleportCostAmount(): Double
 }
